[unrated] Title: 길이에 따른 연산, Time: 0.01 ms, Memory: 4.15 MB -BaekjoonHub

+++ 프로그래머스/unrated/181879. 길이에 따른 연산/길이에 따른 연산.cpp
@@ -3,8 +3,8 @@
 #include <numeric>
 using namespace std;
 
-int solution(vector<int> num_list) {
+long long solution(vector<int> num_list) {
     return num_list.size()<=10?
-        accumulate(num_list.begin(), num_list.end(), 1, multiplies<int>()):
+        accumulate(num_list.begin(), num_list.end(), 1, multiplies<long long>()):
         accumulate(num_list.begin(), num_list.end(), 0);
 }
